requirements.txt
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
DateTime==5.1
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.0
ipython==8.12.2
jedi==0.18.2
jupyter-client==8.2.0
jupyter-core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
zope.interface==6.0
